requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apispec==6.3.0
appdirs==1.4.4
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
aws==0.2.5
Babel==2.14.0
backoff==2.2.1
bandit==1.7.6
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.6.3
boto==2.49.0
boto3==1.34.94
botocore==1.34.94
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
camelot-py==0.11.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.5.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
coverage==7.3.2
crewai==0.30.11
crewai-tools==0.2.6
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.6
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distlib==0.3.4
distro==1.9.0
docker==7.0.0
docstring-parser==0.15
docx2txt==0.8
effdet==0.4.1
elastic-transport==8.10.0
elasticsearch==8.11.0
embedchain==0.1.102
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fabric==3.2.2
Faker==20.1.0
fastapi==0.110.3
fastavro==1.9.4
filelock==3.7.1
filetype==1.2.0
Flask==3.0.0
Flask-APScheduler==1.13.1
flask-babel==4.0.0
Flask-Cors==4.0.0
flask-extend-headers==1.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-QRcode==3.1.0
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.6.0
ghostscript==0.7
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.54.1
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
gptcache==0.1.43
greenlet==3.0.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.59.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==0.5.2
intel-openmp==2021.4.0
invoke==2.2.0
iopath==0.1.10
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.1.20
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.5
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.69
layoutparser==0.3.4
logstash-formatter==0.5.17
lxml==5.2.1
mail-parser==3.15.0
Mako==1.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.1
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.31.0
opencv-python==4.9.0.80
openpyxl==3.1.3
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.3
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.2
parameterized==0.9.0
paramiko==3.4.0
pbr==6.0.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pg8000==1.31.2
pikepdf==9.0.0
Pillow==10.1.0
pillow_heif==0.16.0
pinecone-client==4.1.0
platformdirs==2.5.2
pluggy==1.5.0
polib==1.2.0
portalocker==2.8.2
posthog==3.5.0
prettytable==3.10.0
proto-plus==1.23.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pulsar-client==3.5.0
py==1.11.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
PyGithub==1.59.1
Pygments==2.17.2
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pylance==0.9.18
pylint==3.0.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyNaCl==1.5.0
pypandoc==1.13
pyparsing==3.1.1
pypdf==3.17.4
pypdfium2==4.30.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.366
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.2.2
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==3.0.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.9.3
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
schema==0.7.7
scikit-learn==1.5.0
scipy==1.11.4
scramp==1.4.5
seaborn==0.13.1
selenium==4.21.0
semver==3.0.2
sentence-transformers==3.0.0
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
stevedore==5.1.0
streamlit==1.35.0
streamlit-chat==0.1.1
svglib==1.5.1
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.5.2
timm==1.0.3
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
twilio==9.0.2
typer==0.9.4
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2023.3
tzlocal==5.2
unstructured==0.14.4
unstructured-client==0.22.0
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
uritools==4.0.2
urllib3==2.0.7
uvicorn==0.30.1
virtualenv==20.15.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webargs==8.3.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.0
xhtml2pdf==0.2.15
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.19.2
